+++ CHANGELOG.md
@@ -76,6 +76,7 @@
 - Quick-search: Shift/Ctrl modifiers to perform lookup at any position instead of only the start of strings is now applied only once at first key pressed, so there is no need to keep pressing the modifier key while typing the other chars.
 - Quick-search: minor UI adjustments.
 - Settings: internal changes at properties for future releases. Upgrading from [1.0.0-beta.8](#100-beta8---2026-01-12) will work seamlessly, but previous versions are not supported (settings will be reset).
+- Helpers: support for long paths (>260 chars) in multiple internal file handling functions.
 ### Removed
 - UI: removed multiple old settings related to background colors and art, which are no longer available and replaced by new features. Note old and new UI will not necessarily match after update.
 ### Fixed